+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Note: versioning follows Semver standard with 3 levels
 
+## Release 1.0.0 [2026-09-14]
+
+### Features
+
+- Add option to add metadata to collections
+- Add option to use metadata schemas from iRODS
+
+### Testing
+
+- Test against iRODS as well
+
+
 ## Release 0.9.0 [2025-12-11]
 
 ### Features